+++ src/core/injections/results_based/techniques/eval_based/eb_enumeration.py
@@ -20,6 +20,7 @@ import sys
 from src.utils import menu
 from src.utils import settings
 
+from src.core.injections.controller import checks
 from src.thirdparty.colorama import Fore, Back, Style, init
 from src.core.injections.results_based.techniques.eval_based import eb_injector
 
@@ -43,6 +44,7 @@ def powershell_version(separator, TAG, prefix, suffix, http_request_method, url,
   try:
     ps_version = "".join(str(p) for p in ps_version).replace(" ", "", 1)[:-1]
     if float(ps_version):
+      settings.PS_ENABLED = True
       if menu.options.verbose:
         print ""
       # Output PowerShell's version number
@@ -53,8 +55,9 @@ def powershell_version(separator, TAG, prefix, suffix, http_request_method, url,
       output_file.write("    (!) The PowerShell's version number is " + ps_version + ".\n")
       output_file.close()
   except ValueError:
-    print Fore.YELLOW + "(^) Warning: Heuristics have failed to identify PowerShell's version, which means that many attack vector may be failed." + Style.RESET_ALL 
-
+    print Fore.YELLOW + "(^) Warning: Heuristics have failed to identify PowerShell's version, which means that some payloads or injection techniques may be failed." + Style.RESET_ALL 
+    settings.PS_ENABLED = False
+    checks.ps_check_failed()
 
 """
 Hostname enumeration
@@ -394,20 +397,21 @@ def single_os_cmd_exec(separator, TAG, prefix, suffix, http_request_method, url,
 Check the defined options
 """
 def do_check(separator, TAG, prefix, suffix, http_request_method, url, vuln_parameter, alter_shell, filename):
+  
   if not menu.options.verbose:
     print ""
 
-  if menu.options.ps_version and settings.TARGET_OS == "win":
+  # Check if PowerShell is enabled.
+  if not menu.options.ps_version and settings.TARGET_OS == "win":
+    checks.ps_check()
+
+  if menu.options.ps_version and settings.TARGET_OS == "win" and settings.PS_ENABLED == None:
     powershell_version(separator, TAG, prefix, suffix, http_request_method, url, vuln_parameter, alter_shell, filename)
     settings.ENUMERATION_DONE = True
 
   if menu.options.hostname:
     hostname(separator, TAG, prefix, suffix, http_request_method, url, vuln_parameter, alter_shell, filename)
     settings.ENUMERATION_DONE = True
-  # elif settings.ENUMERATION_DONE == False:
-  #     print ""
-  # else:
-  #     print ""
     
   if menu.options.current_user:
     current_user(separator, TAG, prefix, suffix, http_request_method, url, vuln_parameter, alter_shell, filename)